Support updating from an existing Template
This gives a handy RollBack feature where we can re-use a Template by
it's Object Key on S3. Typically, deployments are Parameter updates
which is both easy to revert and hard to remember what to revert to, but
storing the old Templates could be useful for non-Parameter changes such
as IAM or container resources limits.