Temporarily point services at Heroku datastores
GitHub refuses to see the DNS update so webhook delivery is still going
to the AWS services. By pointing it at Heroku DB/Redis we can keep the
work and data all happening there and keep a single App service up for
now to act as a forwarder. We'll give a little more time and see if it
eventually moves over.