Move RESTYLERS_VERSION from SSM to Stack Parameter
We have to restart services anyway, so having it driven by SSM is
actually making a change two-step. Also, RESTYLER_IMAGE is basically the
same kind of setting, so having the be different isn't great.

We may eventually build SSM-driven deploys, by listening to SSM events
and restarting services on change, but we'd want to do that consistently
across all "runtime" settings.