Separate IAM Stack
The idea is a work-in-progress, but:

- Dedicated Stack for IAM resources
- Can be deployed via a more tightly-controlled CloudGenesis instance
- Can be the only deployment instance with IAM Capabilities
- Creates policies that are either broad (and safe to be) or restricted
  to conventionally-named Resources (restyled/$Environment/*)
- Other Stacks reference the policies by name, can be deployed via less
  managed CG instances without IAM Capabilities (TODO)

One open question is IAM Policies that are currently scoped to a
specific Resource. This is the most secure way to scope IAM, but
requires the IAM Policy be defined in the same Stack (meaning it now
requires IAM Capabilities). This is a direct trade-off and I'm not sure
which is better. We're leaving such cases alone for now, which is why we
can't actually move to the dedicated deployers and removing IAM
Capabilities where possible.