Move sites/ under prod/
This is sort of true in that we may wish to deploy non-prod instances of
these sites. But mostly this is so we can retain the current stacks
because we're unable to do a graceful replacement (the CNAME on the
Distribution is a blocker) and taking down docs will break restyles
looking for the manifests.

Maybe we shouldn't depend on docs for manifests...