Add second restyle-machines Stack
This will be the experimental Stack we'll start deploying the Restyling
Agent changes to. The currently-deployed Stack will be left alone until
we're ready to cut over.

On deploy, this will look like a minor scale up to the system, but I
will immediately deploy the new agent user-data (to only this Stack),
which will get it back out of production rotation, but ready to accept
experimental restyles.